requirements.txt
Django==4.1.7
django-environ==0.9.0
django-split-settings==1.2.0
django_debug_toolbar==3.8.1
pytest==7.2.1
pytest-django==4.5.2
whitenoise==6.3.0
